add PersonData.java

+++ src/test/java/scanner/PersonData.java
@@ -12,7 +12,7 @@ public class PersonData {
         System.out.print("Input age: ");
         int age = in.nextInt();
         System.out.print("Input height (m): ");
-        double height = in.nextDouble();
+        float height = in.nextFloat();
         System.out.print("Input weight (kg): ");
         float weight = in.nextFloat();
         System.out.printf("Name: %s   Age: %d   Height: %.2f m   Weight: %.2f kg\n", name, age, height, weight);
